# output/api-contracts-updated.yaml
openapi: 3.0.0
info:
  title: FormFit Coach API
  description: |
    RESTful API for FormFit Coach platform - connecting fitness enthusiasts with expert trainers for personalized video form analysis.

    ## Authentication
    All authenticated endpoints require either:
    - Session cookie (for web app)
    - JWT token in Authorization header: `Bearer <token>`

    ## Rate Limiting
    Rate limits vary by endpoint and authentication status. Limits are returned in headers:
    - `X-RateLimit-Limit`: Maximum requests allowed
    - `X-RateLimit-Remaining`: Requests remaining
    - `X-RateLimit-Reset`: Unix timestamp when limit resets

    ## Versioning
    API is versioned via URL path (e.g., /api/v1/). Version v1 is current stable version.

  version: 1.0.0
  contact:
    name: FormFit Coach API Support
    email: api-support@formfit.com
servers:
  - url: https://formfit.com/api/v1
    description: Production server
  - url: https://staging.formfit.com/api/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
  schemas:
...
    Comment:
      type: object
      properties:
        id:
          type: string
        feedbackId:
          type: string
        userId:
          type: string
        content:
          type: string
        timestamp:
          type: integer
          description: Video timestamp in seconds
          nullable: true
        createdAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
...
paths:
...
  /auth/verify/{token}:
    get:
      tags:
        - Authentication
      summary: Verify email address
      description: Verifies email using token from verification email
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Email verified successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  redirect:
                    type: string
                    nullable: true
        '400':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/forgot-password:
    post:
      tags:
        - Authentication
      summary: Request password reset
      description: Sends password reset email with time-limited token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Password reset email sent (always returns 200 to prevent email enumeration)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/reset-password:
    post:
      tags:
        - Authentication
      summary: Reset password using token
      description: Resets user password with token from reset email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
                - newPassword
              properties:
                token:
                  type: string
                newPassword:
                  type: string
                  minLength: 8
                  description: Must contain uppercase, lowercase, number, and special character
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '400':
          description: Invalid token or password requirements not met
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /feedback/{id}/comments:
    post:
      tags:
        - Comments
      summary: Add timestamped comment
      description: Add a timestamped comment to feedback (trainer only)
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Feedback ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - timestamp
                - comment
              properties:
                timestamp:
                  type: integer
                  description: Video timestamp in seconds
                  minimum: 0
                comment:
                  type: string
                  minLength: 1
                  maxLength: 500
      responses:
        '201':
          description: Comment added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  comment:
                    $ref: '#/components/schemas/Comment'
                  totalComments:
                    type: integer
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
        '403':
          description: Not assigned trainer for this feedback
  /feedback/{feedbackId}/comments/{commentId}:
    delete:
      tags:
        - Comments
      summary: Delete timestamped comment
      description: Delete a comment (only before feedback submission)
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: feedbackId
          in: path
          required: true
          schema:
            type: string
        - name: commentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Comment deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  remainingComments:
                    type: integer
        '400':
          description: Cannot delete after feedback completed
        '401':
          description: Unauthorized
        '403':
          description: Not comment author
  /credits/checkout:
    post:
      tags:
        - Credits
      summary: Create checkout session
      description: Create Stripe Checkout session for credit package purchase
      security:
        - BearerAuth: []
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - packageId
              properties:
                packageId:
                  type: string
                  enum: [starter, pro, expert]
                successUrl:
                  type: string
                  format: uri
                cancelUrl:
                  type: string
                  format: uri
      responses:
        '200':
          description: Checkout session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId:
                    type: string
                  url:
                    type: string
                    format: uri
        '401':
          description: Unauthorized
        '429':
          description: Rate limit exceeded (max 10 per 5 minutes)
...